adds comicskindom
