+++ src/all/comicskingdom/src/eu/kanade/tachiyomi/extension/all/comicskingdom/ComicsKingdomFactory.kt
@@ -1,0 +1,8 @@
+package eu.kanade.tachiyomi.extension.all.comicskingdom
+
+import eu.kanade.tachiyomi.source.Source
+import eu.kanade.tachiyomi.source.SourceFactory
+
+class ComicsKingdomFactory : SourceFactory {
+    override fun createSources(): List<Source> = listOf(ComicsKingdom("en"), ComicsKingdom("es"))
+}
